chore(edit): move function to utils
