+++ next-tavla/src/Admin/scenarios/Edit/components/SelectAuthorities/utils/removeDuplicatesMultiSelect.ts
@@ -1,1 +1,11 @@
+import { NormalizedDropdownItemType } from '@entur/dropdown'
+
+export function removeDuplicates(data: NormalizedDropdownItemType[]) {
+    const unique: NormalizedDropdownItemType[] = []
+    data.forEach((item) => {
         if (!unique.some((uniqueItem) => uniqueItem.value === item.value)) {
+            unique.push(item)
+        }
+    })
+    return unique
+}
